Move all preprocessor editors from owpreprocess.py
