+++ orangecontrib/spectroscopy/tests/test_owpreprocess.py
@@ -7,7 +7,8 @@ from Orange.preprocess.preprocess import Preprocess
 from orangecontrib.spectroscopy.data import getx
 from orangecontrib.spectroscopy.tests import spectral_preprocess
 from orangecontrib.spectroscopy.tests.spectral_preprocess import pack_editor, wait_for_preview
-from orangecontrib.spectroscopy.widgets.owpreprocess import OWPreprocess, \
+from orangecontrib.spectroscopy.widgets.owpreprocess import OWPreprocess
+from orangecontrib.spectroscopy.widgets.preprocessors.misc import \
     CutEditor, SavitzkyGolayFilteringEditor
 from orangecontrib.spectroscopy.widgets.preprocessors.registry import preprocess_editors
 from orangecontrib.spectroscopy.widgets.preprocessors.utils import BaseEditorOrange, \
