+++ orangecontrib/spectroscopy/widgets/preprocessors/__init__.py
@@ -1,4 +1,5 @@
 # load and register editors
+from . import misc
 from . import als
 from . import atm_corr
 from . import baseline
